Uploaded homework three - focus on File I/O and BFS
Homework three focused on reading data about mansion rooms from a file, and searching each of the mansion rooms for clues. Each room was connected by a doorway, and the constraint was that you had to move from rooms that were connected (pointers!). A Breadth-first search is used to solve this problem.