Uploaded homework six - I/O and sorting algorithm
Homework six involved reading from a file of newspaper writers and assigning them to articles from a file of articles. Each writer had three story types that they preferred, and my algorithm was such that they would never be assigned a story type that didn't match up with a preference. The most difficult was a sorting algorithm, which proved to be an interesting challenge.